Revert C++ codes to patch 34c4b20 & fix mkstemp race condition
